Make constructor params optional
