+++ mod.ts
@@ -1,2 +1,3 @@
 export * from './src/MessageBoxClient.js'
 export * from './src/PeerPayClient.js'
+export * from './src/types.js'
